web/supabase/requirements.txt
supabase==2.3.5
python-dotenv==1.0.1
geopandas>=0.14.3
shapely>=2.0.2
py3dtiles>=3.0.0
numpy>=1.24.0
requests>=2.31.0
owslib>=0.29.0 